+++ assets/customizer/inc/totalpress-more-options.php
@@ -1,4 +1,4 @@
-<?php /* @version 1.0.23 */
+<?php /* @version 1.0.24 */
 if ( ! defined('ABSPATH')) exit;
 TotalPress_Kirki::add_section( 'totalpress_more_options_section', array(
     'title'       => esc_html__('Extend TotalPress','totalpress'),