+++ assets/customizer/inc/totalpress-typography-section.php
@@ -1,4 +1,4 @@
-<?php /* @version 1.0.23 */
+<?php /* @version 1.0.24 */
 if ( ! defined('ABSPATH')) exit;
 TotalPress_Kirki::add_section( 'totalpress_typography_section', array(
     'title'       => esc_html__('Typography','totalpress'),
@@ -10,17 +10,17 @@ TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 	'type'        => 'typography',
 	'settings'    => 'totalpress_theme_body_font',
 	'label'       => esc_html__('Text','totalpress' ),
-	'tooltip'     => esc_html__('Change the font family of your theme. Default is sans-serif.','totalpress'),
+	'tooltip'     => esc_html__('Change the font family of your theme.','totalpress'),
 	'section'     => 'totalpress_typography_section',
 	'priority'    => 1,
 	'transport'   => 'auto',
 	'default'     => array(
-		'font-family' => 'sans-serif',
-		'variant'        => 'regular',
-		'font-size'      => '17px',
-		'line-height'    => '1.5',
-		'letter-spacing' => '0',
-		'text-transform' => 'none'
+		'font-family' 		=> 'inherit',
+		'variant'        	=> 'regular',
+		'font-size'      	=> '17px',
+		'line-height'    	=> '1.5',
+		'letter-spacing' 	=> '0',
+		'text-transform' 	=> 'none'
 	),
 	'choices' => array(
 		'fonts' => array(