+++ assets/customizer/js/totalpress-customizer.js
@@ -1,4 +1,4 @@
-/* @version 1.0.23 */
+/* @version 1.0.24 */
 ( function( $ ) {
 // Header
 	// Site title and description.