+++ assets/js/totalpress-app.js
@@ -1,4 +1,4 @@
-/* @version 1.0.23 */
+/* @version 1.0.24 */
 jQuery(document).foundation();
 	jQuery('input[type="submit"]').addClass('button');
 	jQuery('iframe[src*="youtube.com"],iframe[src*="vimeo.com"],iframe[src*="dailymotion.com"],iframe[src*="videopress.com"]').wrap('<div class="responsive-embed widescreen" />');
@@ -13,8 +13,5 @@ jQuery(document).foundation();
 		event.preventDefault();
 		jQuery('html, body').animate({scrollTop: 0},500);
 	});
+	jQuery(".accordion p:empty, .accordion br, .orbit p:empty, .orbit br").remove();
 	jQuery('.gallery').addClass('clearfix');
-	jQuery('.menu-social-container a').each(function()
-		{
-	  	jQuery(this).wrapInner('<span class="screen-reader-text"></span>');
-	});;